requirements.txt
crispy-bootstrap5==0.6
cryptography==38.0.4
cssselect2==0.7.0
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.9.0
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
urllib3==1.26.13
weasyprint==57.2
webencodings==0.5.1
whitenoise==6.2.0
